Extract `ColumnAddition` tests in a separate class
